供应商

+++ src/main/java/com/chisong/green/farm/app/entity/Supplier.java
@@ -67,7 +67,7 @@ public class Supplier implements Serializable {
     /**
     * 创建人
      */
-    private  Integer creatorId;
+    private  Long creatorId;
 
     /**
     * 公司大图
